Emotion-Detection-For-Facial-Expressions
The project is about emotion detection by measuring facial features using OpenCV(Python). This system can predict and recognize the expression of the person facing the camera. The system has the ability to detect any face and based on training data it can recognize and register the expression and different facial features. It is very simple and e…